+++ packages/std/src/lib.rs
@@ -104,8 +104,8 @@ pub mod testing {
     };
     #[cfg(feature = "stargate")]
     pub use crate::mock::{
-        mock_ibc_channel, mock_ibc_channel_close, mock_ibc_channel_connect, mock_ibc_channel_open,
-        mock_ibc_packet, mock_ibc_packet_ack, mock_ibc_packet_recv, mock_ibc_packet_timeout,
+        mock_ibc_channel_close, mock_ibc_channel_connect, mock_ibc_channel_open,
+        mock_ibc_packet_ack, mock_ibc_packet_recv, mock_ibc_packet_timeout,
     };
 }